Fix conversion warning

+++ Polyhedron/demo/Polyhedron/Plugins/Classification/Classification_plugin.cpp
@@ -835,7 +835,7 @@ public Q_SLOTS:
       if (label_buttons.size() > 1)
         for (int i = position + 1; i < label_buttons.size(); ++ i)
         {
-          int position = i - 1;
+          int position = static_cast<int>(i - 1);
           int x = position / 3;
           int y = position % 3;
